webgateway/pyproject.toml
[tool.poetry]
name = "webgateway"
version = "0.1.0"
description = ""
authors = ["ketov-x <ketov-x@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-ninja = "^0.19.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.5.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
